requirements.txt
streamlit
pyaudio
wave
openai
google.generetiveai
python-dotenv